one_d_two_d_test re configurated (#28)
