+++ hhnk_threedi_plugin/gui/input_data.py
@@ -29,6 +29,7 @@ from hhnk_threedi_plugin.qgis_interaction import load_layers_interaction
 from hhnk_threedi_tools.qgis.get_working_paths import get_working_paths
 from hhnk_threedi_tools.qgis.environment import testEnvironment
 
+
 def setupUi(input_data):
     input_data.setWindowTitle("Model sqlite checks")
     input_data.setMinimumWidth(500)
@@ -63,7 +64,7 @@ def setupUi(input_data):
         file_dialog_title="Selecteer DEM raster (.tif)",
         file_mode=QFileDialog.ExistingFile,
         name_filter="*.tif",    )
-    input_data.dem_selector.setEnabled(False)
+    input_data.dem_selector.setEnabled(True)
 
     input_data.datachecker_label = QLabel("Selecteer datachecker output:")
     input_data.datachecker_selector = fileWidget(
@@ -139,6 +140,8 @@ def setupUi(input_data):
     main_layout.addWidget(input_data.start_sqlite_tests_btn, alignment=Qt.AlignRight)
     input_data.setLayout(main_layout)
 
+
+
 class inputDataDialog(QDialog):
 
     def __init__(self, caller, parent):
@@ -146,6 +149,8 @@ class inputDataDialog(QDialog):
         setupUi(self)
         self.caller = caller
         self.setup_main_paths_signals()
+        self.model_selector.fileSelected.connect(self.caller.reset_ui)
+
         
     def setup_main_paths_signals(self):
         """
@@ -159,7 +164,9 @@ class inputDataDialog(QDialog):
         self.channel_from_profiles_selector.fileSelected.connect(lambda path: self.caller.update_current_paths(channels_shapefile=path))
         self.model_selector.fileSelected.connect(lambda path: self.caller.update_current_paths(model=path))
         self.dem_selector.fileSelected.connect(lambda path: self.caller.update_current_paths(dem=path))
-        self.output_selector.fileSelected.connect(lambda path: self.caller.update_current_paths(sqlite_output=path))
+        # self.output_selector_sqlite.fileSelected.connect(lambda path: self.caller.update_current_paths(sqlite_output=path))
+        # self.result_selector_1d2d.fileSelected.connect(lambda path: self.caller.update_current_paths(one_d_results=path))
+
 
     def set_current_paths(self):
         """
@@ -176,3 +183,20 @@ class inputDataDialog(QDialog):
             self.damo_selector.setFilePath(paths["damo"])
             self.polder_shape_selector.setFilePath(paths["polder_shapefile"])
             self.output_selector.setFilePath(paths["sqlite_tests_output"])
+            # self.one_d_two_d.dem_selector.setFilePath(paths["dem"])
+            
+            
+            # self.output_selector.setFilePath(paths["sqlite_tests_output"])
+
+
+    # def set_current_paths_1d_2d(self):
+    #     """
+    #     Sets current paths as known to main widget to this widget's fields
+    #     """
+    #     paths = self.caller.current_source_paths
+    #     if paths is not None:
+    #         self.model_selector.setFilePath(paths["model"])
+    #         self.dem_selector.setFilePath(paths["dem"])
+    #         self.results_dir_selector.setFilePath(paths["1d2d_results_dir"])
+    #         self.output_selector.setFilePath(paths["1d2d_output"])
+    #         #         E:\02.modellen\model_test_v2_juan
